% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{build_linear}
\alias{build_linear}
\title{Build linear components}
\usage{
build_linear(formula, data = NULL)
}
\arguments{
\item{formula}{A formula specifying a linear predictor.}

\item{data}{A data.frame containign the variables in the formula.}
}
\value{
A list containing the \code{model.matrix} \code{X} and the zero Penalty
matrix \code{P}.
}
\description{
Build linear components
}
