% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{build_nl}
\alias{build_nl}
\title{Build non linear components}
\usage{
build_nl(object, data = NULL)
}
\arguments{
\item{object}{A character vector representing a call to \code{\link[=nl]{nl()}} function.}

\item{data}{A data.frame containign the variables in the formula for \code{\link[=nl]{nl()}}.}
}
\value{
An object created by \code{\link[=nl]{nl()}} function.
}
\description{
Build non linear components
}
