% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{build_ridge}
\alias{build_ridge}
\title{Build ridge components}
\usage{
build_ridge(object, data = NULL)
}
\arguments{
\item{object}{A character vector representing a call to \code{\link[=ridge]{ridge()}} function.}

\item{data}{A data.frame containign the variables in the formula for \code{\link[=ridge]{ridge()}}.}
}
\value{
An object created by \code{\link[=ridge]{ridge()}} function.
}
\description{
Build ridge components
}
