% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{build_smooth}
\alias{build_smooth}
\title{Build smooth components}
\usage{
build_smooth(smooth_term, data)
}
\arguments{
\item{smooth_term}{A smooth term specified by \code{\link[mgcv:s]{mgcv::s()}}.}

\item{data}{A data.frame containign the variables to build the smooth components.}
}
\value{
A list containing the \code{model.matrix} \code{X} and the penalty matrix \code{P}.
}
\description{
Build smooth components
}
