% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib.R
\name{check_derivatives}
\alias{check_derivatives}
\title{Compare Analytical and Numerical Derivatives of the Log-Likelihood}
\usage{
check_derivatives(distrib, x, par, round_digits = 5)
}
\arguments{
\item{distrib}{An object of class \code{\link{distrib}}.}

\item{x}{A numeric vector of observations at which the derivatives are evaluated.}

\item{par}{A list containing the values of the model parameters.}

\item{round_digits}{An integer specifying the number of decimal places
to which the difference between analytical and numerical derivatives is rounded.}
}
\value{
A list with two elements:
\describe{
\item{\code{gradient}}{A data frame comparing the analytical and numerical gradients.}
\item{\code{hessian}}{A data frame comparing the analytical and numerical Hessians
(diagonal and upper off-diagonal elements only).}
}
}
\description{
Evaluates and compares the analytical (explicit) and numerical gradients
and Hessians of the log-likelihood function for a given distribution.
}
\examples{
check_derivatives(normal1(), rnorm(100), par = list(0, 1))
check_derivatives(normal2(), rnorm(100), par = list(0, 1))
check_derivatives(normal3(), rnorm(100), par = list(0, 1))
}
