% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{classify_term}
\alias{classify_term}
\title{Classify terms in formula}
\usage{
classify_term(x)
}
\arguments{
\item{x}{A character vector indicating a formula term.}
}
\value{
The type of the formula term indicated.
}
\description{
Classify terms in formula
}
