% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib.R
\name{gamma1}
\alias{gamma1}
\title{Gamma distribution (\eqn{\mu, \sigma^2} parametrization)}
\usage{
gamma1(link_mu = "log", link_sigma2 = "log")
}
\arguments{
\item{link_mu}{Character string naming the link for the mean parameter
(default \code{"log"}).}

\item{link_sigma2}{Character string naming the link for the variance parameter
(default \code{"log"}).}
}
\value{
An object of class \strong{\code{distrib}}.
See \link{distrib} for the structure common to all distribution objects..
}
\description{
Builds a \code{distrib} object representing a Gamma distribution with mean
\eqn{\mu} and variance \eqn{\sigma^2}.
}
\details{
The \code{gamma1()} function generates a \code{distrib} object representing a Gamma distribution with parameters:
\deqn{\bm{\theta} = (\mu, \sigma^2)}
and pdf:
\deqn{f(y) = \dfrac{\left(\dfrac{\mu}{\sigma^2}\right)^{\frac{\mu^2}{\sigma^2}} y ^ {\frac{\mu^2}{\sigma^2}-1} \exp\left\{ - \dfrac{\mu}{\sigma^2}y\right\}}{\Gamma\left(\dfrac{\mu^2}{\sigma^2}\right)}}
}
\seealso{
\link{distrib} for the structure common to all distribution objects.
}
