% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm.R
\name{gdrm_add_map_coef}
\alias{gdrm_add_map_coef}
\title{Adds mapped coefficients to a gdrm model}
\usage{
gdrm_add_map_coef(mod_comp, map_functions)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}

\item{map_functions}{A list of functions to map parameters from constrained to real line created by \code{\link[=make_map_function]{make_map_function()}}.}
}
\value{
The \code{mod_comp} passed as input with added the \code{par_map} object to each component of the model parameters.
}
\description{
Adds mapped coefficients to a gdrm model
}
