% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm_extractors.R
\name{gdrm_coef}
\alias{gdrm_coef}
\title{Coefficient of gdrm model}
\usage{
gdrm_coef(mod_comp)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}
}
\value{
A list containing the coefficients for each component of the model.
}
\description{
Coefficient of gdrm model
}
