% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm_extractors.R
\name{gdrm_coef_names}
\alias{gdrm_coef_names}
\title{Coefficient names of gdrm model}
\usage{
gdrm_coef_names(mod_comp)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}
}
\value{
A character vector containing names of the parameters.
}
\description{
Coefficient names of gdrm model
}
