% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm.R
\name{gdrm_coef_vector}
\alias{gdrm_coef_vector}
\title{Extract coefficient of gdrm model in a numerical vector}
\usage{
gdrm_coef_vector(mod_comp)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}
}
\value{
A numeric vector containing the coefficients of the gdrm model.
}
\description{
Extract coefficient of gdrm model in a numerical vector
}
