% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm.R
\name{gdrm_eta2_beta2}
\alias{gdrm_eta2_beta2}
\title{Second derivative of eta wrt beta of gdrm model}
\usage{
gdrm_eta2_beta2(mod_comp)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}
}
\value{
A list containing the second derivatives of eta wrt beta for each component of the model.
}
\description{
Second derivative of eta wrt beta of gdrm model
}
