% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm.R
\name{gdrm_eta_beta}
\alias{gdrm_eta_beta}
\title{Derivatives of eta wrt beta of gdrm model}
\usage{
gdrm_eta_beta(mod_comp)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}
}
\value{
A list containing the the derivatives of eta wrt beta for each component of the model.
}
\description{
Derivatives of eta wrt beta of gdrm model
}
