% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm_extractors.R
\name{gdrm_fitted}
\alias{gdrm_fitted}
\title{Predicted values on parameter scale}
\usage{
gdrm_fitted(mod_comp, distrib)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}

\item{distrib}{A \verb{[distrib]} object.}
}
\value{
A list of predicted values on parameter scale.
}
\description{
Predicted values on parameter scale
}
