% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm.R
\name{gdrm_grad}
\alias{gdrm_grad}
\title{Gradient of loglikelihood of gdrm model}
\usage{
gdrm_grad(response, distrib, mod_comp, sum = TRUE)
}
\arguments{
\item{response}{Response variable.}

\item{distrib}{A \verb{[distrib]} object.}

\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}

\item{sum}{Logical. If \code{TRUE} (default) the gradient is return, else if \code{FALSE} the individual contributions to gradient are returned.}
}
\description{
Gradient of loglikelihood of gdrm model
}
