% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm.R
\name{gdrm_grad_map}
\alias{gdrm_grad_map}
\title{Mapped Gradient of loglikelihood of gdrm model}
\usage{
gdrm_grad_map(
  response,
  distrib,
  mod_comp,
  map_functions,
  sum = TRUE,
  penalty = TRUE
)
}
\arguments{
\item{response}{Response variable.}

\item{distrib}{A \verb{[distrib]} object.}

\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}

\item{map_functions}{A list of functions to map parameters from constrained to real line created by \code{\link[=make_map_function]{make_map_function()}}.}

\item{sum}{Logical. If \code{TRUE} (default) the gradient is return, else if \code{FALSE} the individual contributions to gradient are returned.}

\item{penalty}{Logical. If \code{TRUE} (default) the penalized gradient is returned.}
}
\value{
Mapped Gradient of loglikelihood function.
}
\description{
Mapped Gradient of loglikelihood of gdrm model
}
