% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm_loglik_derivatives.R
\name{gdrm_hessian_map}
\alias{gdrm_hessian_map}
\title{Mapped Hessian of loglikelihood of gdrm model}
\usage{
gdrm_hessian_map(
  response,
  distrib,
  mod_comp,
  map_functions,
  sum = TRUE,
  penalty = TRUE,
  P = NULL,
  expected = TRUE,
  predict = NULL,
  fitted = NULL,
  g = NULL,
  hess = NULL
)
}
\arguments{
\item{response}{Response variable.}

\item{distrib}{A \verb{[distrib]} object.}

\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}

\item{map_functions}{A list of functions to map parameters from constrained to real line created by \code{\link[=make_map_function]{make_map_function()}}.}

\item{sum}{Logical. If \code{TRUE} (default) the hessian is returned, else if \code{FALSE} the individual contributions to the hessian are returned.}

\item{penalty}{Logical. If \code{TRUE} (default) the penalized hessian is returned.}

\item{P}{Optional. An object returned by \verb{[gdrm_penalty()]}.}

\item{expected}{Logical. If \code{TRUE} (default) use expected Fisher Information if available.}

\item{predict}{Optional. An object returned by \verb{[gdrm_predict()]}.}

\item{fitted}{Optional. An object returned by \verb{[gdrm_fitted()]}.}

\item{g}{Optional. An object returned by \verb{[gdrm_grad()]} with \code{penalty = FALSE} and \code{sum} argument as specified in the call of this function.}

\item{hess}{Optional. An object returned by \verb{[gdrm_hessian()]} with \code{penalty = FALSE} and \code{sum} argument as specified in the call of this function.}
}
\value{
Mapped Hessian of loglikelihood function.
}
\description{
Mapped Hessian of loglikelihood of gdrm model
}
