% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm_extractors.R
\name{gdrm_hyperpar}
\alias{gdrm_hyperpar}
\title{Hyperparameters of gdrm model}
\usage{
gdrm_hyperpar(mod_comp)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}
}
\value{
A list containing the hyperparameters for each component of the model.
}
\description{
Hyperparameters of gdrm model
}
