% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm_extractors.R
\name{gdrm_hyperpar_vector}
\alias{gdrm_hyperpar_vector}
\title{Extract hyperparameters of gdrm model in a numerical vector}
\usage{
gdrm_hyperpar_vector(mod_comp)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}
}
\value{
A numeric vector containing the hyperparameters of the gdrm model.
}
\description{
Extract hyperparameters of gdrm model in a numerical vector
}
