% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{gdrm_invert_coef}
\alias{gdrm_invert_coef}
\title{Invert map transformation for gdrm coefficients}
\usage{
gdrm_invert_coef(par, coef_id, map_functions)
}
\arguments{
\item{par}{A vector containing mapped coefficients to be inverted.}

\item{coef_id}{A vector of ids which assigns each coefficients in par to the respective model parameter.}

\item{map_functions}{A list of functions to map parameters from constrained to real line created by \code{\link[=make_map_function]{make_map_function()}}.}
}
\value{
A vector containing the coefficient of the model inverted to their original constrained space.
}
\description{
Invert map transformation for gdrm coefficients
}
