% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm_loglik_derivatives.R
\name{gdrm_l2_theta2}
\alias{gdrm_l2_theta2}
\title{Hessian of the loglikelihood function wrt theta of gdrm model}
\usage{
gdrm_l2_theta2(response, mod_comp, distrib, expected = TRUE)
}
\arguments{
\item{response}{Response variable.}

\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}

\item{distrib}{A \verb{[distrib]} object.}

\item{expected}{Logical. If \code{TRUE} (default) use expected Fisher Information if available.}
}
\value{
A list containing the hessian of the loglikelihood wrt eta for each model parameter.
}
\description{
Hessian of the loglikelihood function wrt theta of gdrm model
}
