% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm.R
\name{gdrm_l_theta}
\alias{gdrm_l_theta}
\title{Derivatives of the loglikelihood function wrt theta of gdrm model}
\usage{
gdrm_l_theta(response, mod_comp, distrib)
}
\arguments{
\item{response}{Response variable.}

\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}

\item{distrib}{A \verb{[distrib]} object.}
}
\value{
A list containing the the derivatives of the loglikelihood wrt eta for each model parameter.
}
\description{
Derivatives of the loglikelihood function wrt theta of gdrm model
}
