% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm.R
\name{gdrm_loglik}
\alias{gdrm_loglik}
\title{Loglikelihood function of gdrm model}
\usage{
gdrm_loglik(response, distrib, mod_comp, penalty = TRUE)
}
\arguments{
\item{response}{Response variable.}

\item{distrib}{A \verb{[distrib]} object.}

\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}

\item{penalty}{Logical. If \code{TRUE} (default) the penalty term is added to the loglikelihood.}
}
\value{
A numeric value which is the loglikelihood of the gdrm model.
}
\description{
Loglikelihood function of gdrm model
}
