% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{gdrm_map_coef}
\alias{gdrm_map_coef}
\title{Map coefficients of the model to unconstrained space}
\usage{
gdrm_map_coef(mod_comp, map_functions)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}

\item{map_functions}{A list of functions to map parameters from constrained to real line created by \code{\link[=make_map_function]{make_map_function()}}.}
}
\value{
A vector containing mapped coefficients of the gdrm model.
}
\description{
Map coefficients of the model to unconstrained space
}
