% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm_extractors.R
\name{gdrm_penalty}
\alias{gdrm_penalty}
\title{Build penalty matrix for gdrm model}
\usage{
gdrm_penalty(mod_comp, hyperpar = TRUE)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}

\item{hyperpar}{Logical. If \code{TRUE} (default), the elements of the penalty matrix are multiplied by the respective hyperparameter.}
}
\value{
A list of penalty matrices for each model parameter.
}
\description{
Build penalty matrix for gdrm model
}
