% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm_extractors.R
\name{gdrm_predict}
\alias{gdrm_predict}
\title{Predicted values on link scale}
\usage{
gdrm_predict(mod_comp)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}
}
\value{
A list of predicted values on link scale.
}
\description{
Predicted values on link scale
}
