% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm_fitters.R
\name{gdrm_sa}
\alias{gdrm_sa}
\title{Simulated annealing optimization for gdrm}
\usage{
gdrm_sa(
  response,
  mod_comp,
  distrib,
  map_functions,
  sa_control_list = sa_control()
)
}
\arguments{
\item{response}{Response variable.}

\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}

\item{distrib}{A \code{\link{distrib}} object.}

\item{map_functions}{A list of functions to map parameters from constrained to real line created by \code{\link[=make_map_function]{make_map_function()}}.}

\item{sa_control_list}{A list of options returned by \code{\link[=sa_control]{sa_control()}} function.}
}
\value{
A list containing:
\item{\code{mod_comp}}{\code{mod_comp} object updated with new parameters value.}
\item{\code{history}}{Numeric vector containing the loglikelihood values at each iterations.}
}
\description{
Simulated annealing optimization for gdrm
}
