% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{gdrm_start}
\alias{gdrm_start}
\title{Starting values of coefficients for gdrm models}
\usage{
gdrm_start(response, mod_comp, distrib, coef_bounds)
}
\arguments{
\item{response}{Response variable.}

\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}

\item{distrib}{A \verb{[distrib]} object specifying the distribution assumed for the response variable.}

\item{coef_bounds}{A list of coefficient bounds. The list should contain an element for each model parameters, and any of those element should be another list containing named elements being numerical vector of lower and upper bounds for parameters.}
}
\value{
Update \code{mod_comp} setting starting values for the coefficients.
}
\description{
Starting values of coefficients for gdrm models
}
