% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm.R
\name{gdrm_update_coef}
\alias{gdrm_update_coef}
\title{Update parameters of a gdrm model}
\usage{
gdrm_update_coef(par, mod_comp)
}
\arguments{
\item{par}{A numeric vector of updated coefficients of a gdrm model.}

\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}
}
\value{
A new \code{mod_comp} object with updated parameters.
}
\description{
Update parameters of a gdrm model
}
