% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{gdrm_update_hyperpar}
\alias{gdrm_update_hyperpar}
\title{Update hyperparameters of a gdrm model}
\usage{
gdrm_update_hyperpar(par, mod_comp)
}
\arguments{
\item{par}{A numeric vector of updated hyperparameters of a gdrm model.}

\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}
}
\value{
A new \code{mod_comp} object with updated hyperparameters.
}
\description{
Update hyperparameters of a gdrm model
}
