% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_coef_names}
\alias{get_coef_names}
\title{Extract coefficient names}
\usage{
get_coef_names(mod_comp)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}
}
\value{
A list containing the coefficient names for each model parameters.
}
\description{
Extract coefficient names
}
