% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{get_response}
\alias{get_response}
\title{Extract response variable from a formula}
\usage{
get_response(formulae, data)
}
\arguments{
\item{formulae}{A \code{formulae} object, which consists of formulas for different parameters of the model separated by \code{&} operator.}

\item{data}{A data.frame from which the response variable has to be extracted.}
}
\value{
A vector containing the response variable.
}
\description{
Extract response variable from a formula
}
