% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{interpret_formulae}
\alias{interpret_formulae}
\title{Interpret a formulae}
\usage{
interpret_formulae(formulae, distrib, data)
}
\arguments{
\item{formulae}{A formula with different parts related to different parameters separated by \code{&} operator.}

\item{distrib}{An object of class \link{distrib}.}

\item{data}{A data.frame containign the variables in the formula.}
}
\description{
Interpret a formulae
}
