% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_coef_bounds}
\alias{make_coef_bounds}
\title{Create list of coefficient bounds}
\usage{
make_coef_bounds(coef_bounds, coef_names)
}
\arguments{
\item{coef_bounds}{A list of coefficient bounds. The list should contain an element for each model parameters, and any of those element should be another list containing named elements being numerical vector of lower and upper bounds for parameters.}

\item{coef_names}{A list of coefficient names created by \code{\link[=get_coef_names]{get_coef_names()}} function.}
}
\value{
A list of coefficient bounds to pass to \code{\link[=make_map_function]{make_map_function()}} function.
}
\description{
Create list of coefficient bounds
}
