% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_link.R
\name{make_link}
\alias{make_link}
\title{Create a link function object for gdrm}
\usage{
make_link(link)
}
\arguments{
\item{link}{Character string specifying the link function. Options include:
\itemize{
\item \code{"logit"}: log(theta/(1-theta))
\item \code{"probit"}: inverse of the standard normal CDF
\item \code{"cauchit"}: inverse of the Cauchy CDF
\item \code{"cloglog"}: log(-log(1-theta))
\item \code{"identity"}: no transformation
\item \code{"log"}: natural logarithm
\item \code{"sqrt"}: square root
\item \code{"1/theta^2"}: inverse square
\item \code{"inverse"}: reciprocal
}}
}
\value{
A list with S3 class \code{"link-gnlm"} containing the following components:
\itemize{
\item \code{linkfun}: Function that transforms the theta to eta
\item \code{linkinv}: Function that transforms eta to theta
\item \code{theta.eta}: Function returning the derivative of theta with respect to eta (\eqn{d\theta/d\eta})
\item \code{theta2.eta2}: Function returning the second derivative of theta with respect to eta (\eqn{d^2\theta/d^2\eta^2})
\item \code{valideta}: Function that checks if the eta values are valid
\item \code{name}: Character string naming the link function
}
}
\description{
This function creates a link function object similar to \code{stats::make.link()} but extends it by
including the second derivative of theta with respect to eta (\code{theta2.eta2}).
}
\note{
This function extends \code{stats::make.link()} by including the \code{theta2.eta2} component,
which calculates the second derivative of theta with respect to eta.
}
