% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl.R
\name{nl}
\alias{nl}
\title{Specification of non linear component}
\usage{
nl(formula, parameters, data = NULL)
}
\arguments{
\item{formula}{A right-sided formula specifying a non linear expression.}

\item{parameters}{A list with elements of the same name of the parameters specifying a vector of lower and upper bounds.}

\item{data}{A data.frame containing variables used to compute the non linear function.}
}
\value{
A list with tree elements:
\describe{
\item{\code{f}}{A function with argument the parameters \code{par} to compute the non linear expression.}
\item{\code{jac}}{A function with argument the parameters \code{par} to compute the jacobian of \code{f}.}
\item{\code{hes}}{A function with argument the parameters \code{par} to compute the hessian of \code{f}.}
}
}
\description{
Specification of non linear component
}
\examples{
formula <- ~ theta0 * (x1^theta1) * (x2^theta2)
parameters <- list(theta0 = c(0, Inf), theta1 = c(0, Inf), theta2 = c(0, Inf))
data <- abs(data.frame(x1 = rnorm(5), x2 = rnorm(5)))
r <- nl(formula, parameters, data)
par <- list(1, 2, 3)
r$f(par)
r$jac(par)
r$hes(par)
}
