% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib.R
\name{normal1}
\alias{normal1}
\title{Normal distribution (\eqn{\mu, \sigma^2} parametrization)}
\usage{
normal1(link_mu = "identity", link_sigma2 = "log")
}
\arguments{
\item{link_mu}{Character string naming the link for the mean parameter
(default \code{"identity"}).}

\item{link_sigma2}{Character string naming the link for the variance parameter
(default \code{"log"}).}
}
\value{
An object of class \strong{\code{distrib}}.
See \link{distrib} for the structure common to all distribution objects..
}
\description{
Builds a \code{distrib} object representing a Normal distribution with mean
\eqn{\mu} and variance \eqn{\sigma^2}.
}
\details{
The \code{normal1()} function generate a \code{distrib} object representing a normal distribution with parameters:
\deqn{\bm{\theta} = (\mu, \sigma^2)}
and pdf:
\deqn{f(y) = \dfrac{1}{\sqrt{2 \pi \sigma^2}} \exp \left\{-\dfrac{(y - \mu)^2}{2\sigma^2}\right\}}
}
\seealso{
\link{distrib} for the structure common to all distribution objects.
}
