% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib.R
\name{normal2}
\alias{normal2}
\title{Normal distribution (\eqn{\mu, \sigma} parametrization)}
\usage{
normal2(link_mu = "identity", link_sigma = "log")
}
\arguments{
\item{link_mu}{Character string naming the link for the mean parameter
(default \code{"identity"}).}

\item{link_sigma}{Character string naming the link for the standard deviation parameter
(default \code{"log"}).}
}
\value{
An object of class \strong{\code{distrib}}.
See \link{distrib} for the structure common to all distribution objects..
}
\description{
Builds a \code{distrib} object representing a Normal distribution with mean
\eqn{\mu} and standard deviation \eqn{\sigma}.
}
\details{
The \code{normal2()} function generate a \code{distrib} object representing a normal distribution with parameters:
\deqn{\bm{\theta} = (\mu, \sigma)}
and pdf:
\deqn{f(y) = \dfrac{1}{\sqrt{2 \pi} \sigma} \exp \left\{-\dfrac{1}{2}\left(\dfrac{y - \mu}{\sigma}\right)^2\right\}}
}
\seealso{
\link{distrib} for the structure common to all distribution objects.
}
