% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib.R
\name{normal3}
\alias{normal3}
\title{Normal distribution (\eqn{\mu, \tau^2} parametrization)}
\usage{
normal3(link_mu = "identity", link_tau2 = "log")
}
\arguments{
\item{link_mu}{Character string naming the link for the mean parameter
(default \code{"identity"}).}

\item{link_tau2}{Character string naming the link for the precision parameter
(default \code{"log"}).}
}
\value{
An object of class \strong{\code{distrib}}.
See \link{distrib} for the structure common to all distribution objects..
}
\description{
Builds a \code{distrib} object representing a Normal distribution with mean
\eqn{\mu} and precision \eqn{\tau^2}.
}
\details{
The \code{normal3()} function generate a \code{distrib} object representing a normal distribution with parameters:
\deqn{\bm{\theta} = (\mu, \tau^2)}
and pdf:
\deqn{f(y) = \sqrt{\dfrac{\tau^2}{2 \pi}} \exp \left\{-\dfrac{\tau^2 (y - \mu)^2}{2}\right\}}
}
\seealso{
\link{distrib} for the structure common to all distribution objects.
}
