% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib.R
\name{print.distrib}
\alias{print.distrib}
\title{Print method for distrib objects}
\usage{
\method{print}{distrib}(x, ...)
}
\arguments{
\item{x}{A \code{distrib} object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for distrib objects
}
