% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm.R
\name{print.gdrm}
\alias{print.gdrm}
\title{Print method for gdrm objects}
\usage{
\method{print}{gdrm}(x, ...)
}
\arguments{
\item{x}{A gdrm object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for gdrm objects
}
