% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm_fitters.R
\name{sa_control}
\alias{sa_control}
\title{Control options for Simulated Annealing algorithm}
\usage{
sa_control(
  iter = 30000,
  t_start = 100,
  t_end = 1,
  sd = 0.1,
  save_history = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{iter}{Integer. Number of iterations.}

\item{t_start}{Numeric. Temperature at the start of algorithm.}

\item{t_end}{Numeric. Temperature at the end of algorithm.}

\item{sd}{Numeric. Standard deviation used for sampling new parameters value.}

\item{save_history}{Logical. If \code{TRUE} (default) loglikelihood values at each iterations are stored and returned.}

\item{verbose}{Logical. If \code{TRUE} (default) progress is printed.}
}
\value{
A list containing Simulated annealing options.
}
\description{
Control options for Simulated Annealing algorithm
}
