% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm.R
\name{gdrm_control}
\alias{gdrm_control}
\title{Control options for gdrm()}
\usage{
gdrm_control(force_intercept = TRUE)
}
\arguments{
\item{force_intercept}{Logical. If \code{TRUE} (default) a global intercept is forced to the model if none linear part is present.}
}
\value{
A list of class \code{gdrm_control}.
}
\description{
Control options for gdrm()
}
