% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm_extractors.R
\name{gdrm_fitted}
\alias{gdrm_fitted}
\title{Fitted values on parameter scale}
\usage{
gdrm_fitted(mod_comp, distrib, predict = NULL)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}

\item{distrib}{A \verb{[distrib]} object.}

\item{predict}{Optional. An object returned by \verb{[gdrm_predict()]}.}
}
\value{
A list of fitted values on parameter scale.
}
\description{
Fitted values on parameter scale
}
