% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm.R
\name{gdrm_theta2_eta2}
\alias{gdrm_theta2_eta2}
\title{Second derivative of theta wrt eta of gdrm model}
\usage{
gdrm_theta2_eta2(mod_comp, distrib)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}

\item{distrib}{A \verb{[distrib]} object.}
}
\value{
A list containing the second derivative of theta wrt eta for each model parameter.
}
\description{
Second derivative of theta wrt eta of gdrm model
}
