% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrm_loglik_derivatives.R
\name{gdrm_theta_eta}
\alias{gdrm_theta_eta}
\title{First derivative of theta wrt eta of gdrm model}
\usage{
gdrm_theta_eta(mod_comp, distrib)
}
\arguments{
\item{mod_comp}{A list of model components from \verb{[interpret_formulae()]}.}

\item{distrib}{A \verb{[distrib]} object.}
}
\value{
A list containing the first derivative of theta wrt eta for each model parameter.
}
\description{
First derivative of theta wrt eta of gdrm model
}
