% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{interpret_formulae}
\alias{interpret_formulae}
\title{Interpret a formulae}
\usage{
interpret_formulae(formulae, distrib, data, force_intercept = TRUE)
}
\arguments{
\item{formulae}{A \code{formulae} object, which consists of formulas for different parameters of the model separated by \code{&} operator.}

\item{distrib}{An object of class \link{distrib}.}

\item{data}{A data.frame containign the variables in the formula.}

\item{force_intercept}{Logical. If \code{TRUE} (default) a global intercept is forced to the parameter formula if none linear part is present. See \code{\link[=interpret_formula]{interpret_formula()}} for more details.}
}
\value{
A list for each parameter of the distribution containing model matrices and penalty matrices
for each term of the formula
}
\description{
Interpret a formulae
}
